refactor: check ktlint
